# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: TriReservas API
  description: |-
    TriReservas es una aplicación web para la gestión de alojamientos y reservas, con roles diferenciados para 
    usuarios y anfitriones. Desarrollada con Spring Boot, JPA y Angular, ofrece una solución moderna y eficiente para la
    administración de estancias.
    
    Requisitos Funcionales:
    * [Aquí](https://docs.google.com/spreadsheets/d/14pBo766cm17JRFVoYanh_tZuF3VmeBVc5iNh8Avp4dE/edit?usp=sharing) podrá 
    encontrar un apoyo en hojas de cálculo para la identificación de requisitos funcionales y el progreso del proyecto.

    Autores: 
    * Juan Sebastián Londoño Ramírez [github](https://github.com/Juan3bitas)
    * Angie Carolina Grajales Giraldo [github](https://github.com/angiec-grajalesg)
    * Juan Diego García Nieto [github](https://github.com/Juan-Devgo)
  version: 1.1.1
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
...
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Crear un nuevo usuario
      description: Crea un nuevo usuario en el sistema, cumpliendo con el requisito REQ-RYA-U-01 y REQ-RYA-A-08.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: Usuario creado satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Usuario creado satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "RYA01-1", message: "Error en la solicitud al tratar de crear un usuario."}
        '409':
          description: Conflicto - Usuario ya existe.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Conflicto - Usuario ya existe."
                  content:
                    default: {code: "RYA01-2", message: "El usuario ya se encuentra registrado."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-05", message: "Error inesperado al tratar de crear un usuario."}
  /api/auth/user/login/{mode}:
    post:
      tags:
        - Auth
      summary: Autenticar usuario y obtener token JWT
      description: Autentica un usuario y genera un token JWT para la autenticación en futuras solicitudes, cumpliendo con el requisito REQ-RYA-U-01, REQ-RYA-H-01, REQ-RYA-A-08.
      operationId: loginUser
      parameters:
        - name: mode
          in: path
          required: true
          description: Modo de autenticación, puede ser por 'email' o 'phone'.
          schema:
            type: string
            enum: [email, phone]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Autenticación exitosa, token JWT generado."
                  content:
                    type: string
                    description: Token JWT para autenticación
        '400':
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Credenciales inválidas."
                  content:
                    default: {code: "RYA01-1", message: "Error en la solitud al tratar de iniciar sesión."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-06", message: "Error inesperado al tratar de iniciar sesión."}
  /api/auth/host/login/{mode}:
    post:
      tags:
        - Auth
      summary: Autenticar anfitrión y obtener token JWT
      description: Autentica un anfitrión y genera un token JWT para la autenticación en futuras solicitudes, cumpliendo con el requisito REQ-RYA-H-01 y REQ-RYA-A-08.
      operationId: loginHost
      parameters:
        - name: mode
          in: path
          required: true
          description: Modo de autenticación, puede ser por 'email' o 'phone'.
          schema:
            type: string
            enum: [ email, phone ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Autenticación exitosa, token JWT generado."
                  content:
                    type: string
                    description: Token JWT para autenticación
        '400':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Credenciales inválidas."
                  content:
                    default: {code: "RYA08-2", message: "Error en la solitud al tratar de iniciar sesión."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-08", message: "Error inesperado al tratar de iniciar sesión."}
  /api/auth/reset_password:
    post:
      tags:
        - Auth
      summary: Solicitar restablecimiento de contraseña
      description: Envía un correo electrónico con un enlace para restablecer la contraseña del usuario, cumpliendo con el requisito REQ-RYA-A-09.
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Solicitud de restablecimiento de contraseña enviada satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Solicitud de restablecimiento de contraseña enviada satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "GTU06-1", message: "Error en la solicitud al tratar de solicitar el restablecimiento de contraseña."}
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Usuario no encontrado."
                  content:
                    default: {code: "GTU06-2", message: "No se encontró el usuario especificado para solicitar el restablecimiento de contraseña."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-09", message: "Error inesperado al tratar de solicitar el restablecimiento de contraseña."}
    patch:
      tags:
        - Auth
      summary: Actualizar contraseña de usuario que lo solicitó anteriormente
      description: Actualiza la contraseña de un usuario que solicitó el cambio anteriormente.
      operationId: updatePasswordWithCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Contraseña restablecida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Contraseña restablecida satisfactoriamente."
                  content:
                    default: null
        '422':
          description: Error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error de validación."
                  content:
                    default: {code: "GTU07-1", message: "El código actual enviado no coincide."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-24", message: "Error inesperado al tratar de restablecer la contraseña"}
...
  /api/image:
    post:
      tags:
        - Image
      summary: crear una imagen en la nube
      description: Crea una imagen en la nube se devuelve su enlace
      operationId: postImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              default: null # TODO
      responses:
        '200':
          description: Imagen posteada en la nube correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Imagen posteada en la nube correctamente"
                  content:
                    default: null # TODO
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-24", message: "Error inesperado al tratar de postear la imagen en la nube."}
    delete:
      tags:
        - Image
      summary: Eliminar una imagen de la nube
      description: Elimina una imagen que había sido posteada en la nube anteriormente.
      operationId: deleteImage
      responses:
        '200':
          description: Imagen eliminada de la nube correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Imagen eliminada de la nube correctamente"
                  content:
                    default: null # TODO
        default:
          description: Error Inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-25", message: "Error inesperado al tratar de eliminar la imagen de la nube."}
components:
  schemas:
    UserDTO:
      description: Esquema que representa un usuario para transferencia de datos.
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          format: number
        role:
          type: string
          enum: [USER, HOST]
        birthDate:
          type: string
          format: date
      required:
        - name
        - email
        - role
        - birthDate
    UserLoginDTO:
      description: Esquema que representa las credenciales de inicio de sesión de un usuario.
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          format: password
      required:
        - password
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT